+++ src/DatePointTest.php
@@ -17,7 +17,7 @@ use DateTimeImmutable;
 use DateTimeInterface;
 use PHPUnit\Framework\TestCase;
 
-final class DatepointTest extends TestCase
+final class DatePointTest extends TestCase
 {
     private string $timezone;
 
@@ -36,7 +36,7 @@ final class DatepointTest extends TestCase
      */
     public function testIsAfter(Period $interval, DateTimeInterface $input, bool $expected): void
     {
-        self::assertSame($expected, Datepoint::fromDate($input)->isAfter($interval));
+        self::assertSame($expected, DatePoint::fromDate($input)->isAfter($interval));
     }
 
     public function isAfterProvider(): array
@@ -80,7 +80,7 @@ final class DatepointTest extends TestCase
      */
     public function testIsBefore(Period $interval, DateTimeInterface $input, bool $expected): void
     {
-        self::assertSame($expected, Datepoint::fromDate($input)->isBefore($interval));
+        self::assertSame($expected, DatePoint::fromDate($input)->isBefore($interval));
     }
 
     public function isBeforeProvider(): array
@@ -111,7 +111,7 @@ final class DatepointTest extends TestCase
 
     public function testDatepointBorderingOn(): void
     {
-        $datepoint = Datepoint::fromDateString('2018-01-18 10:00:00');
+        $datepoint = DatePoint::fromDateString('2018-01-18 10:00:00');
         $duration = Duration::fromDateString('3 minutes');
 
         $intervalBorderOnStartTrue = Period::after($datepoint, $duration, Period::EXCLUDE_START_INCLUDE_END);
@@ -138,7 +138,7 @@ final class DatepointTest extends TestCase
      */
     public function testIsDuring(Period $interval, $input, bool $expected): void
     {
-        $datepoint = $input instanceof DateTimeInterface ? Datepoint::fromDate($input) : Datepoint::fromDateString($input);
+        $datepoint = $input instanceof DateTimeInterface ? DatePoint::fromDate($input) : DatePoint::fromDateString($input);
 
         self::assertSame($expected, $datepoint->isDuring($interval));
     }
@@ -199,7 +199,7 @@ final class DatepointTest extends TestCase
      */
     public function testStarts(Period $interval, DateTimeInterface $index, bool $expected): void
     {
-        self::assertSame($expected, Datepoint::fromDate($index)->isStarting($interval));
+        self::assertSame($expected, DatePoint::fromDate($index)->isStarting($interval));
     }
 
     public function startsDataProvider(): array
@@ -225,7 +225,7 @@ final class DatepointTest extends TestCase
      */
     public function testFinishes(Period $interval, DateTimeInterface $index, bool $expected): void
     {
-        self::assertSame($expected, Datepoint::fromDate($index)->isEnding($interval));
+        self::assertSame($expected, DatePoint::fromDate($index)->isEnding($interval));
     }
 
     public function isEndingDataProvider(): array